# cloud/backend/mon-projet-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mon API Laravel 2026",
        "description": "Documentation de mon API générée avec Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Local"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login utilisateur",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mdp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie"
                    },
                    "401": {
                        "description": "Email ou mot de passe invalide"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un nouvel utilisateur",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "mdp",
                                    "Id_role"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "mdp": {
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "Id_role": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fire_user_id": {
                                        "type": "string",
                                        "example": "b7Hk6pV0bVQq..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/entreprises": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Lister les entreprises",
                "operationId": "7f9edf2d966d33903c18bd90ee7a0dfd",
                "responses": {
                    "200": {
                        "description": "Liste des entreprises"
                    }
                }
            },
            "post": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Créer une entreprise",
                "operationId": "883d1fed0716732a97e0d428ec084d0c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Feno SARL"
                                    },
                                    "is_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Entreprise créée"
                    }
                }
            }
        },
        "/api/entreprises/{id}": {
            "get": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Afficher une entreprise",
                "operationId": "fa454084a3a74ba55fe3634872a81a17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entreprise trouvée"
                    }
                }
            },
            "put": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Modifier une entreprise",
                "operationId": "db09dd91c137fe5832ebfdc3ca6fe5f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entreprise modifiée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Entreprises"
                ],
                "summary": "Supprimer une entreprise",
                "operationId": "fc502310c692349bea4b024f2f81207e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entreprise supprimée"
                    }
                }
            }
        },
        "/api/problemes": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Lister les problèmes",
                "operationId": "39ac43e54d22eddd8c10e7156202b543",
                "responses": {
                    "200": {
                        "description": "Liste des problèmes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Créer un problème",
                "operationId": "00fbb512747f8ff79b35e475824e4d52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Id_entreprise",
                                    "Id_signalement"
                                ],
                                "properties": {
                                    "budget": {
                                        "type": "number",
                                        "example": 15000.5
                                    },
                                    "surface": {
                                        "type": "number",
                                        "example": 120.75
                                    },
                                    "is_deleted": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "Id_entreprise": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Id_signalement": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Problème créé"
                    }
                }
            }
        },
        "/api/problemes/{id}": {
            "get": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Afficher un problème",
                "operationId": "ecc098749ca4c7e0c385fed6ebf2ca75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problème trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Modifier un problème",
                "operationId": "e6bf25526188ee8e8058a43265141da1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problème modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Problemes"
                ],
                "summary": "Supprimer un problème",
                "operationId": "550b1db8cf9ebadc6652b85962541b4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problème supprimé"
                    }
                }
            }
        },
...
    },
...
}